==> Data-Engineering-MLOps-Exam-Assignment/4_model_deployment.ipynb <==
# Deploy model

#!pip install -U hopsworks --quiet

import warnings
warnings.filterwarnings("ignore")

import hopsworks

project = hopsworks.login()

fs = project.get_feature_store()

## <span style="color:#ff5f27;"> Deploy model</span>

%%writefile predict_example.py
import os
import numpy as np
import hsfs
import joblib

class Predict(object):

    def __init__(self):
        """ Initializes the serving state, reads a trained model"""
        # load the trained model
        self.model = joblib.load(os.environ["ARTIFACT_FILES_PATH"] + "/xgboost.pkl")
        print("Initialization Complete")

    def predict(self, inputs):
        """ Serves a prediction request using a trained model"""
        # Convert inputs to numpy array to handle batch predictions
        inputs_array = np.array(inputs)
        # Check if the input is a single instance or batch
        if inputs_array.ndim == 1:
            # Reshape if a single instance to match input shape expected by the model
            inputs_array = inputs_array.reshape(1, -1)
        # Make predictions
        return self.model.predict(inputs_array).tolist()  # Return predictions as list

import os
dataset_api = project.get_dataset_api()

uploaded_file_path = dataset_api.upload("predict_example.py", "Models", overwrite=True)
predictor_script_path = os.path.join("/Projects", project.name, uploaded_file_path)

print("Uploaded file path:", uploaded_file_path)
print("Predictor script path:", predictor_script_path)


## Create the deployment

# Use the model name from the previous notebook.
mr = project.get_model_registry()
model = mr.get_model("BeerRatingPredictionModel1", version=1)

# Create and start the deployment
deployment = model.deploy(
    name="beerratingmodeldeployment1",  # Give it a descriptive name
    serving_tool="KSERVE",  # Specify the serving tool
    script_file=predictor_script_path  # The path to the prediction script
)

print("Deployment: " + deployment.name)
deployment.describe()
deployment.start()
